# contracts/openapi/v1/inframeld-v1.json
{
  "components": {
    "schemas": {
      "CurrentSessionResponse": {
        "properties": {
          "principalId": {
            "format": "uuid",
            "title": "Principalid",
            "type": "string"
          }
        },
        "required": [
          "principalId"
        ],
        "title": "CurrentSessionResponse",
        "type": "object"
      },
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "Represent an Inframeld HTTP error using RFC 9457 Problem Details.\n\nAttributes:\n    type: Absolute problem URI; about:blank is used for generic HTTP errors.\n    title: Short public summary without occurrence-specific internal data.\n    status: Error status in the range 400 through 599.\n    detail: Safe public explanation selected by the HTTP adapter.\n    code: Stable public code; clients must tolerate unfamiliar future codes.\n    request_id: Correlation UUID serialized as requestId, never authorization.\n    errors: Optional bounded collection of already sanitized request issues.\n\nThe response builder enforces the selected definition's type/code/status\npairing and whether errors is appropriate. This model does not map Python\nexceptions, authenticate callers, authorize retries, or render tracebacks.\nServer-side extra fields are rejected; client parsers must still tolerate\nfuture RFC extension members.",
        "properties": {
          "code": {
            "description": "Stable Inframeld code paired with the problem type.",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "description": "Reviewed explanation without internal or submitted sensitive values.",
            "maxLength": 512,
            "minLength": 1,
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "At most 20 sanitized issues; the list may be incomplete.",
            "title": "Errors"
          },
          "operationId": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Original operation identity for an in-progress idempotency retry.",
            "title": "Operationid"
          },
          "requestId": {
            "description": "Server request identifier, also returned in X-Request-ID.",
            "format": "uuid",
            "title": "Requestid",
            "type": "string"
          },
          "status": {
            "description": "HTTP error status, equal to the actual response status.",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Stable human-readable summary of the problem type.",
            "maxLength": 128,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Primary RFC 9457 problem-type URI.",
            "format": "uri",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "ProjectUseOnlyGrantRequest": {
        "properties": {
          "actionId": {
            "const": "create-access-groups",
            "title": "Actionid",
            "type": "string"
          },
          "expectedAccessRevision": {
            "minimum": 0.0,
            "title": "Expectedaccessrevision",
            "type": "integer"
          },
          "recipientPrincipalId": {
            "format": "uuid",
            "title": "Recipientprincipalid",
            "type": "string"
          }
        },
        "required": [
          "recipientPrincipalId",
          "actionId",
          "expectedAccessRevision"
        ],
        "title": "ProjectUseOnlyGrantRequest",
        "type": "object"
      },
      "ProjectUseOnlyGrantResponse": {
        "description": "Describe the original change - current grants require a fresh authorized read.",
        "properties": {
          "accessRevision": {
            "title": "Accessrevision",
            "type": "integer"
          },
          "actionId": {
            "const": "create-access-groups",
            "title": "Actionid",
            "type": "string"
          },
          "canGrant": {
            "const": false,
            "default": false,
            "title": "Cangrant",
            "type": "boolean"
          },
          "idempotencyExpiresAt": {
            "format": "date-time",
            "title": "Idempotencyexpiresat",
            "type": "string"
          },
          "operationId": {
            "title": "Operationid",
            "type": "string"
          },
          "projectId": {
            "format": "uuid",
            "title": "Projectid",
            "type": "string"
          },
          "recipientPrincipalId": {
            "format": "uuid",
            "title": "Recipientprincipalid",
            "type": "string"
          }
        },
        "required": [
          "operationId",
          "projectId",
          "recipientPrincipalId",
          "actionId",
          "accessRevision",
          "idempotencyExpiresAt"
        ],
        "title": "ProjectUseOnlyGrantResponse",
        "type": "object"
      },
      "ValidationIssue": {
        "additionalProperties": false,
        "description": "Describe one sanitized issue in an incoming request.\n\nAttributes:\n    location: Request component containing the issue.\n    path: Verified public field names or array indices, relative to that\n        component. An empty path identifies the component as a whole.\n    code: Product-owned validation category, independent of library names.\n    message: Reviewed corrective description without submitted values.\n\nThis model checks the output shape, not whether a string contains a secret.\nThe validation translator must sanitize locations and messages first.",
        "properties": {
          "code": {
            "description": "Stable Inframeld validation category.",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Code",
            "type": "string"
          },
          "location": {
            "description": "Request component containing the issue.",
            "enum": [
              "body",
              "path",
              "query",
              "header",
              "cookie",
              "request"
            ],
            "title": "Location",
            "type": "string"
          },
          "message": {
            "description": "Safe corrective description without submitted values.",
            "maxLength": 256,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "path": {
            "description": "Verified public fields and indices relative to the request component.",
            "items": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "minimum": 0.0,
                  "type": "integer"
                }
              ]
            },
            "maxItems": 8,
            "title": "Path",
            "type": "array"
          }
        },
        "required": [
          "location",
          "path",
          "code",
          "message"
        ],
        "title": "ValidationIssue",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HumanBearer": {
        "description": "Opaque human Hydra access token. Application keys do not authenticate a human.",
        "scheme": "bearer",
        "type": "http"
      },
      "KratosSession": {
        "description": "Kratos browser session. Cookie-authenticated writes also require a trusted Origin and X-Inframeld-CSRF: 1.",
        "in": "cookie",
        "name": "ory_kratos_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Governed retrieval and release workflows.",
    "title": "Inframeld API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "getHealth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Gethealth",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Health",
        "tags": [
          "health"
        ]
      }
    },
    "/v1/projects/{project_id}/grants": {
      "post": {
        "operationId": "assignProjectUseOnlyGrant",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": ".*\\S.*",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUseOnlyGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectUseOnlyGrantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid authentication request",
            "headers": {
              "WWW-Authenticate": {
                "description": "Trusted Bearer challenge when applicable.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "description": "Trusted Bearer challenge when applicable.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Access denied"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict: stale_revision, idempotency_key_reused, idempotency_in_progress, or conflict."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "KratosSession": []
          },
          {
            "HumanBearer": []
          }
        ],
        "summary": "Assign Use Only",
        "tags": [
          "access"
        ]
      }
    },
    "/v1/session": {
      "get": {
        "operationId": "getCurrentSession",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid authentication request",
            "headers": {
              "WWW-Authenticate": {
                "description": "Trusted Bearer challenge when applicable.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "description": "Trusted Bearer challenge when applicable.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Access denied"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "KratosSession": []
          },
          {
            "HumanBearer": []
          }
        ],
        "summary": "Current Session",
        "tags": [
          "access"
        ]
      }
    }
  }
}